+++ ketos/data_handling/database_interface.py
@@ -29,6 +29,7 @@
     This module provides functions to create and use HDF5 databases as storage for acoustic data. 
 """
 
+import librosa
 import tables
 import os
 import ast
@@ -37,7 +38,8 @@ import numpy as np
 from ketos.utils import tostring
 from ketos.audio_processing.audio import AudioSignal
 from ketos.audio_processing.spectrogram import Spectrogram, MagSpectrogram, PowerSpectrogram, CQTSpectrogram, MelSpectrogram, ensure_same_length
-from ketos.data_handling.data_handling import find_wave_files, AnnotationTableReader, rel_path_unix
+from ketos.data_handling.data_handling import find_wave_files, AnnotationTableReader, rel_path_unix, SpecProvider
+from ketos.data_handling.parsing import SpectrogramConfiguration
 from tqdm import tqdm
 from sys import getsizeof
 from psutil import virtual_memory
@@ -304,6 +306,9 @@ def write_spec(table, spec, id=None):
         id_str = ''
     else:
         id_str = id
+
+    # spectrogram offset is not stored, so shift annotations
+    spec._shift_annotations(delay=-spec.tmin)
 
     if spec.labels is not None:
         labels_str = tostring(spec.labels, decimals=0)
@@ -670,12 +675,11 @@ def parse_boxes(boxes):
     
     return parsed_boxes
 
-def create_spec_database(output_file, input_dir, annotations_file=None,\
+def create_spec_database(output_file, input_dir, annotations_file=None, spec_config=None,\
         sampling_rate=None, channel=0, window_size=0.2, step_size=0.02, duration=None,\
-        flow=None, fhigh=None, max_size=1E9, progress_bar=False, verbose=True, cqt=False,\
+        overlap=0, flow=None, fhigh=None, max_size=1E9, progress_bar=False, verbose=True, cqt=False,\
         bins_per_octave=32, **kwargs):
     """ Create a database with magnitude spectrograms computed from raw audio (*.wav) files
-    
         
         One spectrogram is created for each audio file using either a short-time Fourier transform (STFT) or
         a constant-Q transform (CQT).
@@ -697,6 +701,11 @@ def create_spec_database(output_file, input_dir, annotations_file=None,\
         The internal file structure of the database file will mirror the structure of the 
         data directory where the audio data is stored. See the example below.
 
+        Note that if spec_config is specified, the following arguments are ignored: 
+        sampling_rate, window_size, step_size, duration, overlap, flow, fhigh, cqt, bins_per_octave.
+
+        TODO: Modify implementation so that arguments are not ignored when spec_config is specified.
+
         Args:
             output_file: str
                 Full path to output database file (*.h5)
@@ -704,6 +713,8 @@ def create_spec_database(output_file, input_dir, annotations_file=None,\
                 Full path to folder containing the input audio files (*.wav)
             annotations_file: str
                 Full path to file containing annotations (*.csv)
+            spec_config: SpectrogramConfiguration
+                Spectrogram configuration object.
             sampling_rate: float
                 If specified, audio data will be resampled at this rate
             channel: int
@@ -714,6 +725,8 @@ def create_spec_database(output_file, input_dir, annotations_file=None,\
                 Step size (seconds) used for computing the spectrogram
             duration: float
                 Duration in seconds of individual spectrograms.
+            overlap: float
+                Overlap in seconds between consecutive spectrograms.
             flow: float
                 Lower cut on frequency (Hz)
             fhigh: float
@@ -767,83 +780,45 @@ def create_spec_database(output_file, input_dir, annotations_file=None,\
         max_ann = areader.get_max_annotations()
 
     # spectrogram writer
-    swriter = SpecWriter(output_file=output_file, max_size=max_size, max_annotations=max_ann, verbose=verbose)
+    swriter = SpecWriter(output_file=output_file, max_size=max_size, max_annotations=max_ann, verbose=verbose, ignore_wrong_shape=True)
 
-    # get all wav files in the folder
-    files = find_wave_files(path=input_dir, fullpath=True, subdirs=True)
+    if spec_config is None:
+        spec_config = SpectrogramConfiguration(rate=sampling_rate, window_size=window_size, step_size=step_size,\
+            bins_per_octave=bins_per_octave, window_function=None, low_frequency_cut=flow, high_frequency_cut=fhigh,\
+            length=duration, overlap=overlap, type=['Mag', 'CQT'][cqt])
+
+    # spectrogram provider
+    provider = SpecProvider(path=input_dir, channel=channel, spec_config=spec_config)
 
     # subfolder unix structure
+    files = provider.files
     subfolders = list()
     for f in files:
         sf = rel_path_unix(f, input_dir)
         subfolders.append(sf)
 
-    # loop over files
-    for i in tqdm(range(len(files)), disable = not progress_bar):
-    
-        f = files[i]
-        sf = subfolders[i]
+    # loop over files    
+    num_files = len(files)
+    for i in tqdm(range(num_files), disable = not progress_bar):
 
-        # check if files exists
-        exists = os.path.exists(f)
-        if exists is False:
-            continue
-
-        # read audio
-        a = AudioSignal.from_wav(path=f, channel=channel)  
+        # get next spectrogram
+        spec = next(provider)
 
         # add annotations
         if areader is not None:
-            labels, boxes = areader.get_annotations(a.tag)
-            a.annotate(labels, boxes) 
+            labels, boxes = areader.get_annotations(spec.file_dict[0])
+            spec.annotate(labels, boxes) 
 
-        # check spectrogram size (estimated)
-        mem = virtual_memory()
-        siz = getsizeof(a.data) * window_size / step_size
+        # save spectrogram(s) to file        
+        path = subfolders[i] + 'spec'
+        swriter.cd(path)
+        swriter.write(spec)
 
-        # segment, if spectrogram size exceeds 10% of system memory
-        num_segs = int(np.ceil(siz / (0.1 * mem.total)))
-        length = a.duration() / num_segs
-        if duration is not None:
-            length = min(length, duration)
-
-        segs = a.segment(length=length, pad=True, keep_time=True)
-
-        for seg in segs:
-
-            # resample
-            if sampling_rate is not None:
-                seg.resample(new_rate=sampling_rate) 
-
-            # compute the spectrogram
-            if not cqt:
-                s = MagSpectrogram(audio_signal=seg, winlen=window_size, winstep=step_size, decibel=True) 
-            else:
-                s = CQTSpectrogram(audio_signal=seg, winstep=step_size, fmin=flow, fmax=fhigh, bins_per_octave=bins_per_octave, decibel=True)
-
-            # crop frequencies
-            s.crop(flow=flow, fhigh=fhigh) 
-
-            # if duration is not specified, use duration of first spectrogram
-            if duration is None: 
-                d = s.duration()
-            else:
-                d = duration
-
-            # ensure desired duration 
-            specs = s.segment(length=d, pad=True, **kwargs)
-
-            # save spectrogram(s) to file        
-            path = sf + 'spec'
-            swriter.cd(path)
-            for spec in specs:
-                swriter.write(spec)
-
-        # attempt to free up some memory                
-        del segs
-        del a
+    if swriter.num_ignored > 0:
+        print('Ignored {0} spectrograms with wrong shape'.format(swriter.num_ignored))
 
     swriter.close()
+
 
 class SpecWriter():
     """ Saves spectrograms to a database file (*.h5).
@@ -863,6 +838,8 @@ class SpecWriter():
                 The default values is max_size=1E9 (1 Gbyte)
             verbose: bool
                 Print relevant information during execution such as files written to disk
+            ignore_wrong_shape: bool
+                Ignore spectrograms that do not have the same shape as previously saved spectrograms. Default is False.
 
         Attributes:
             base: str
@@ -895,6 +872,12 @@ class SpecWriter():
                     ’w’: Write; a new file is created (an existing file with the same name would be deleted).
                     ’a’: Append; an existing file is opened for reading and writing, and if the file does not exist it is created.
                     ’r+’: It is similar to ‘a’, but the file must already exist.
+            ignore_wrong_shape: bool
+                Ignore spectrograms that do not have the same shape as previously saved spectrograms. Default is False.
+            num_ignore: int
+                Number of ignored spectrograms
+            spec_shape: tuple
+                Spectrogram shape
 
             Example:
 
@@ -923,7 +906,7 @@ class SpecWriter():
                 >>> print(f.root.spec)
                 /spec (Table(3,), fletcher32, shuffle, zlib(1)) ''
     """
-    def __init__(self, output_file, max_size=1E9, verbose=True, max_annotations=100, mode='w'):
+    def __init__(self, output_file, max_size=1E9, verbose=True, max_annotations=100, mode='w', ignore_wrong_shape=False):
         
         self.base = output_file[:output_file.rfind('.')]
         self.ext = output_file[output_file.rfind('.'):]
@@ -934,8 +917,11 @@ class SpecWriter():
         self.path = '/'
         self.name = 'spec'
         self.verbose = verbose
-        self.spec_counter = 0
         self.mode = mode
+        self.ignore_wrong_shape = ignore_wrong_shape
+        self.spec_counter = 0
+        self.num_ignored = 0
+        self.spec_shape = None
 
     def cd(self, fullpath='/'):
         """ Change the current directory within the database file system
@@ -972,14 +958,21 @@ class SpecWriter():
         # open/create table
         tbl = self._open_table(path=path, name=name, shape=spec.image.shape) 
 
-        # write spectrogram to table
-        write_spec(tbl, spec)
-        self.spec_counter += 1
+        if self.spec_counter == 0:
+            self.spec_shape = spec.image.shape
 
-        # close file if size reaches limit
-        siz = self.file.get_filesize()
-        if siz > self.max_file_size:
-            self.close(final=False)
+        # write spectrogram to table
+        if spec.image.shape == self.spec_shape or not self.ignore_wrong_shape:
+            write_spec(tbl, spec)
+            self.spec_counter += 1
+
+            # close file if size reaches limit
+            siz = self.file.get_filesize()
+            if siz > self.max_file_size:
+                self.close(final=False)
+
+        else:
+            self.num_ignored += 1
 
     def close(self, final=True):
         """ Close the currently open database file, if any